feat: 주류 상세 페이지 컴포넌트 수정
